+++ scripts/put-azure-credentials.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+cat >> local.properties << EOL
+AZURE_USERNAME=${1:-$AZURE_USERNAME}
+AZURE_PASSWORD=${2:-$AZURE_PASSWORD}
+EOL