+++ scripts/put-bintray-secrets-gradle.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 cat >> local.properties << EOL
-bintray.user=${1:-$BINTRAY_USER}
-bintray.key=${2:-$BINTRAY_KEY}
-bintray.repo=${3:-$BINTRAY_REPO}
-bintray.user.org=${4:-$BINTRAY_USER_ORG}
+maven.user=${1:-$MAVEN_USER}
+maven.key=${2:-$MAVEN_KEY}
+maven.repoUrl=${3:-$MAVEN_REPO}
+maven.signingKeyId=${4:-$GDP_SIGNING_KEY_ID}
+maven.secretKeyPath=${5:-$GDP_KEY_SECRET_PATH}
+maven.publicKeyPassword=${6:-$GDP_KEY_PASSWORD}
 EOL