+++ sdk/appcenter-crashes/src/main/java/com/microsoft/appcenter/crashes/utils/ErrorLogHelper.java
@@ -139,6 +139,18 @@ public class ErrorLogHelper {
      */
     private static File sPendingMinidumpDirectory;
 
+    /**
+     * Key for saving deviceInfo to JSON.
+     */
+    @VisibleForTesting
+    static String DEVICE_INFO_KEY = "DEVICE_INFO";
+
+    /**
+     * Key for saving userId to JSON.
+     */
+    @VisibleForTesting
+    static String USER_ID_KEY = "USER_ID";
+
     @NonNull
     public static ManagedErrorLog createErrorLog(@NonNull Context context, @NonNull final java.lang.Thread thread, @NonNull final Throwable throwable, @NonNull final Map<java.lang.Thread, StackTraceElement[]> allStackTraces, final long initializeTimestamp) {
         return createErrorLog(context, thread, getModelExceptionFromThrowable(throwable), allStackTraces, initializeTimestamp, true);
@@ -275,6 +287,7 @@ public class ErrorLogHelper {
         File deviceInfoFile = new File(directorySubfolder, ErrorLogHelper.DEVICE_INFO_FILE);
         try {
             Device deviceInfo = DeviceInfoHelper.getDeviceInfo(context);
+            String userIdContext = UserIdContext.getInstance().getUserId();
             deviceInfo.setWrapperSdkName(WRAPPER_SDK_NAME_NDK);
 
             /* To JSON. */
@@ -283,9 +296,12 @@ public class ErrorLogHelper {
             deviceInfo.write(writer);
             writer.endObject();
             String deviceInfoString = writer.toString();
+            JSONObject jsonObject = new JSONObject();
+            jsonObject.put(DEVICE_INFO_KEY, deviceInfoString);
+            jsonObject.put(USER_ID_KEY, userIdContext);
 
             /* Write file. */
-            FileManager.write(deviceInfoFile, deviceInfoString);
+            FileManager.write(deviceInfoFile, jsonObject.toString());
         } catch (DeviceInfoHelper.DeviceInfoException | IOException | JSONException e) {
             AppCenterLog.error(Crashes.LOG_TAG, "Failed to store device info in a minidump folder.", e);
 
@@ -329,13 +345,39 @@ public class ErrorLogHelper {
     }
 
     /**
-     * Look for 'deviceinfo' file inside the minidump folder and parse it.
+     * Get deviceInfo data.
      *
      * @param logFolder folder where to look for stored device information.
      * @return a device information or null.
      */
     @Nullable
     public static Device getStoredDeviceInfo(File logFolder) {
+        String deviceInfoString = getContextInformation(logFolder);
+        if (deviceInfoString == null) {
+            return null;
+        }
+        return parseDevice(deviceInfoString);
+    }
+
+    /**
+     * Get userId data.
+     * @param logFolder folder where to look for stored userId.
+     * @return userId or null.
+     */
+    public static String getStoredUserInfo(File logFolder) {
+        String userInformationString = getContextInformation(logFolder);
+        if (userInformationString == null) {
+            return null;
+        }
+        return parseUserId(userInformationString);
+    }
+
+    /**
+     * Get data about userId and deviceInfo in JSON format.
+     * @param logFolder - path to folder where placed file with data about userId and deviceId.
+     * @return - data about userId and deviceId in JSON format or null.
+     */
+    static String getContextInformation(File logFolder) {
         File[] files = logFolder.listFiles(new FilenameFilter() {
 
             @Override
@@ -348,20 +390,49 @@ public class ErrorLogHelper {
             return null;
         }
         File deviceInfoFile = files[0];
-        String deviceInfoString = FileManager.read(deviceInfoFile);
-        if (deviceInfoString == null) {
+        String contextInfoString = FileManager.read(deviceInfoFile);
+        if (contextInfoString == null) {
             AppCenterLog.error(Crashes.LOG_TAG, "Failed to read stored device info.");
             return null;
         }
-        return parseDevice(deviceInfoString);
+        return contextInfoString;
     }
 
+    /**
+     * Look for 'userId' data in file inside the minidump folder and parse it.
+     * @param contextInformation - data with information about userId.
+     * @return userId or null.
+     */
     @VisibleForTesting
-    static Device parseDevice(String deviceInfoString) {
+    static String parseUserId(String contextInformation) {
+        try {
+            JSONObject jsonObject = new JSONObject(contextInformation);
+            if (jsonObject.has(USER_ID_KEY)) {
+                return jsonObject.getString(USER_ID_KEY);
+            }
+        } catch (JSONException e) {
+            AppCenterLog.error(Crashes.LOG_TAG, "Failed to deserialize user info.", e);
+        }
+        return null;
+    }
+
+    /**
+     * Look for 'deviceInfo' data in file inside the minidump folder and parse it.
+     * @param contextInformation - data with information about userId.
+     * @return deviceInfo or null.
+     */
+    @VisibleForTesting
+    static Device parseDevice(String contextInformation) {
         try {
             Device device = new Device();
-            JSONObject jsonObject = new JSONObject(deviceInfoString);
-            device.read(jsonObject);
+            JSONObject jsonObject = new JSONObject(contextInformation);
+            JSONObject deviceJson;
+            if (jsonObject.has(DEVICE_INFO_KEY)) {
+                deviceJson = new JSONObject(jsonObject.getString(DEVICE_INFO_KEY));
+            } else {
+                deviceJson = jsonObject;
+            }
+            device.read(deviceJson);
             return device;
         } catch (JSONException e) {
             AppCenterLog.error(Crashes.LOG_TAG, "Failed to deserialize device info.", e);
@@ -469,8 +540,7 @@ public class ErrorLogHelper {
         return report;
     }
 
-    @VisibleForTesting
-    static void setErrorLogDirectory(File file) {
+    public static void setErrorLogDirectory(File file) {
         sErrorLogDirectory = file;
     }