+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # App Center SDK for Android Change Log
 
+## Version 4.1.1
+
+### App Center Distribute Play
+
+* **[Fix]** Fix `onNoReleaseAvailable` callback signature.
+
+### App Center Distribute
+
+* **[Fix]** Fix `NullPointerException` occurring when settings dialog was intended to be shown, but there is no foreground activity at that moment.
+* **[Fix]** Fix a crash when download manager application was disabled.
+* **[Fix]** Fix showing the title in the push notification while downloading a new release.
+
+### App Center Crashes
+
+* **[Fix]** Fix formatting of stack trace in the `ErrorReport`.
+* **[Fix]** Fix setting `userId` value in NDK crashes before sending.
+
+___
+
 ## Version 4.1.0
 
 ### App Center Crashes